requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==4.4.2
ffmpeg-python==0.2.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
moviepy==1.0.3
multidict==5.1.0
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pkg-resources==0.0.0
proglog==0.1.9
PyMySQL==1.0.2
requests==2.25.1
SQLAlchemy==1.4.17
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.4.1
